(#10) map, sprite: TypeScript to Kotlin (still broken)
