+++ src/main/kotlin/me/fornever/atoman/sprites/SpriteFrameInfo.kt
@@ -1,0 +1,11 @@
+package me.fornever.atoman.sprites
+
+interface SpriteFrameInfo {
+    frame: {
+        x: number;
+        y: number;
+        w: number;
+        h: number;
+    };
+    duration: number;
+}
