+++ src/main/kotlin/me/fornever/atoman/sprites/SpriteInfo.kt
@@ -1,0 +1,6 @@
+package me.fornever.atoman.sprites
+
+interface SpriteInfo {
+    frames: { [name: string]: SpriteFrameInfo };
+    meta: { size: { w: number; h: number; } }
+}
